Add image cleanup to the docker example commands
The default Docker behavior is to leave the container filesystem around
for later inspection. If doing a lot of little commands (like our
examples), these can really add up.

This change is to add the flag to the example docker invocations so
Docker will clean up the image when finished. This only affects the
temporary images in which the command is executed; not the named images
pulled down from, e.g., GCR.