Implement basic sigaction in enclave
Current implementation supports only sa_handler(int), full support for
sigaction struct with sa_sigaction(int, siginfo_t *, void *) needs to be
provided in follow ups.
Since in simulation mode the frames are unaware of being interrupted,
the sigaction implementation is different for hardware and simulation.
If it's simulation mode, we pass a pointer to a function inside enclave
to host; if it's hardware mode, this pointer will be set to null:
For hardware, sigacation stores a pointer to the user signal handler
inside enclave, and registers |EnterAndHandleSignal| as the signal
handler on the host, which will enter enclave and call the user signal
handler;
For simulation, sigaction also stores a pointer to the user signal
handler inside enclave. On host side, the function registered as signal
handler is a |TranslateToBridgeAndHandleSignal| function, will translate
signum to bridge value and call the |TranslateFromBridgeAndHandleSignal|
function inside enclave, which translates the bridge signum to the
newlib value and calls the user signal handler. Those extra layers are
used to solve the inconsistency of values inside/outside enclave.

PiperOrigin-RevId: 196703139