Use port-autoselection in gRPC tests
This updates the secure and insecure gRPC server tests to use port
auto-selection, which is preferrable to grpc_pick_unused_port_or_die()
because it doesn't require the gRPC port server.

This change required updating the EnclaveServer class to allow callers
of the Run() entry-point to retrieve the server's host and port. The
Run() entry-point now supports two code paths: START_SERVER and
GET_SERVER_ADDRESS.

PiperOrigin-RevId: 197083236