Implement sigprocmask() in enclave
A signal mask (sigset_t) is maintained inside enclave. Everytime a
signal is passed into the enclave, we will check the signal mask and
reject signals that are masked.

When sigprocmask() is called, we separate the signals to block and to
unblock.  We unblock the signals inside enclave first, then pass it to
the host to block/unblock the signals, then we block the signals inside
the enclave. This is to avoid inconsistent behavior when a signal
arrives while sigprocmask() is called.

Translation for signal mask action and sigset_t are provided when across
enclave boundary.

PiperOrigin-RevId: 198797962