Use StatusSerializer to return the signal status
The signal handling ecall was returning int to indicate the possible
error during execution. This commit uses StatusSerializer to output a
status to better describe the possible error, like the other ecall does.
The status propagated out of the enclave is not logged anywhere
currently, since the signal handler is not supposed to grab any locks.
We may want to save it to a buffer and logged by another thread in the
future if it turned out to be important.

PiperOrigin-RevId: 199384984