Refactor gRPC auth bridge code
This change refactors the bridge layer of the gRPC auth stack that is
responsible for converting C++ structures and protobufs into
C-compatible structures. This change adds a conversion function from the
C++ EnclaveCredentialsOptions structure to the C
grpc_enclave_credential_options structure.

The primary motivation behind this CL is to provide conversion functions
that can be used in other parts of the gRPC code, rather than using
one-off conversion functions all over the place. The newly added
function can be used in the end2end test fixture, allowing us to reuse
richer credentials configuration available in the C++ layer in the
end2end tests.

PiperOrigin-RevId: 200436749