Change `enclaves` arg to a dict, add `test_args`
This commit changes enclave_test's `enclave`/`enclaves` arguments from a
list of enclaves to a dictionary of enclave names to enclave targets.
The dictionary is used in combination with an added `test_args` argument
to interpolate each enclave's name as the path to its built binary in
each string in `test_args`. The interpolated arguments are passed as
command-line arguments to the test binary that is built from a cc_test
of the given srcs, data, deps, etc. arguments.

PiperOrigin-RevId: 200797353