Add static map libs to credentials factories deps
This change includes the AssertionGenerator and AssertionVerifier
libraries required for each credentials factory directly in the
dependency list of the credentials library.

Note that although the credentials factories don't make use of these
libraries directly, the libraries are required to be
statically-registered in order for attestation to work correctly. The
Generator and Verifier libraries are marked "alwayslink" so that adding
them to deps will always link them into the target application, even if
they appear unused.

PiperOrigin-RevId: 201092369