Add a test for gRPC between two enclaves
This change adds a new test that runs two enclaves: one hosts a gRPC
server and the other uses a gRPC client to communicate with a gRPC
server. The test verifies that the gRPC client can form a secure channel
to the gRPC server based on SGX local attestation, and make an RPC.

This test will be reused in the future to test additional attestation
mechanisms and gRPC credentials configurations.

PiperOrigin-RevId: 201092697