Move sockaddr_storage to socket.h
sockaddr_storage was begin defined in in.h, because it had visibility
there to sockaddr_in and sockaddr_in6 and could base the size on those.

However, since sockaddr_storage needs to be defined in socket.h, this
resulted in it including in.h, which is a layering inversion and created
a circular include dependency.

Instead, this commit moves sockaddr_storage into socket.h where it
belongs and bases it on a fixed size known to be big enough for
supported families in Asylo (IPv4 and IPv6).

Also changing to using the __sa_family_t type provided by newlib. Doing
so resulted in a smaller type, affecting struct packing.  Adding packed
attribute to ensure sockaddr_in{,6} can be safely cast to sockaddr.

PiperOrigin-RevId: 202392579