Move public symbols out of asylo::sgx namespace
This change moves Asylo symbols that are considered to be public out of
the asylo::sgx namespace and into the top-level asylo namespace. This is
to stay consistent with how other public symbols in Asylo are named.
Unfortunately, this will break backwards compatibility for code that
depended on the old symbols.

PiperOrigin-RevId: 202687883