Fix the flaky signal test
The signal test was flaky in some scenarios because it's trying to
generate an output status message inside the enclave, which calls
`enc_untrusted_malloc` to allocate untrusted memory for the output
message. And when this ocall returns from the host its reentry is
rejected by the enclave since in a signal handler cssa is 1. We
currently allow signal handling ecall to enter enclave with cssa=1, but
not for the ocalls triggered inside the enclave that are reentering the
enclave, so fix this by going back to return an integer from inside the
enclave and interpret it outside the enclave. Also, the status is not
currently used since signal handler is not supposed to obtain a mutex
and cannot do any type of logging.

PiperOrigin-RevId: 204040772