Update Asylo gRPC dependency to v1.13.0
This change updates the gRPC dependency to v1.13.0 and adds a patch file
for that release.

The patch file no longer contains our custom time_asylo.cc and
env_asylo.cc files. Instead, we use the POSIX definitions available in
gRPC.

This also updates the enclave TSI handshaker implementation to work with
recent changes to the TSI handshaker API.

PiperOrigin-RevId: 204510206