Remove thread from enclave_*secure_server tests
This thread was being spawned with no purpose other than to block until
the gRPC shut down.  But, nothing was using it to perform an action once
the server did shut down, so it was adding complexity without value.

This thread also caused occasional test failures due to nondeterministic
cleanup of this thread in relation to the enclave destruction.

PiperOrigin-RevId: 204823980