Remove unneeded RunServer logic from EnclaveServer
This change refactors the EnclaveServer to support a cleaner model for
management of the gRPC server. It removes the blocking RunServer() logic
from the enclave's Run() entry-point, as this logic prevented clean
shutdown. It moves the server initialization to the enclave's
Initialize() method. The enclave's Run() entry-point no longer takes an
input command. Instead, it just writes the server's address to the
output proto.

PiperOrigin-RevId: 204964425