Fix the issue that signal handler clears TLS
The signal handling inside an enclave is implemented through an ecall.
However, an ecall would clear all the thread local storage when it
enters the enclave (unless it is invoked by an ocall). Therefore, when a
signal handler enters an enclave, it clears the thread local storage
unless the signal arrives during an ocall. This was causing signal test
flaky when it uses thread local storage.

This commit fixes the issue by checking the cssa to make sure the ecall
is not interrupting an existing ecall before initializing thread data.

PiperOrigin-RevId: 205008346