Add serialization/deserialization for SigningKey
This change adds a Serialize() method to the SigningKey interface and
provides and implementation in EcdsaP256Sha256SigningKey. It also adds
a new factory method to EcdsaP256Sha256SigningKey that creates a key
from a DER-encoded private key. It also updates the test for
EcdsaP256Sha256SigningKey to print the serialized SigningKey used in
each test so that failed runs are reproducible.

PiperOrigin-RevId: 205731579