Allow _SC_NPROCESSORS_CONF in sysconf(3) calls
gRPC has changed to call sysconf(_SC_NPROCESSORS_CONF), so we should
support it.

Also make sysconf return a long (as specified by POSIX) but use int64_t
in transit across the bridge.

Finally, rename a few variables in existing sysconf(3) code and change
the int_syscall_return field of SyscallsTestOutput from an int32 to an
int64 so that the sysconf test is guaranteed to not suffer from integer
overflow.

PiperOrigin-RevId: 206379523