Update examples folder to use Asylo locally
Previously, the examples folder was meant to serve as an example of how
an external project can use Asylo as a dependency.  However, having a
directory like that sitting inside of the Asylo source created confusion
and logistical issues.

Instead, the source here will integrate with the Asylo tree (with
provisions still in place to make sure it doesn't take dependencies on
the Asylo internals) and the asylo-examples tarball that acts as an
example of using Asylo from an external project will be derived from
(but no longer be identical to) this directory.

PiperOrigin-RevId: 206390928