De-templatize byte_container_util inputs
This change modifies byte_container_util so that the serialization
functions now take a vector of ByteContainerView objects as their input.
Previously, these functions took a vector of containers as their input.
The change also modifies the functions so that they work with arbitrary
byte containers as the output parameter.

PiperOrigin-RevId: 207352762