Allow IOManager dispatchers with non-int returns
CallWithContext and CallWithHandler were written intending to allow the
IOAction template parameter function to return any type. However, in
the error handling, it returned -1, which makes it only able to work
with integer return types.

In this change, the value to return in the case of error is deduced
based on the return type of the IOaction function. For signed integer
return types, -1 continues to be chosen.  However, for pointer types,
nullptr is chosen as the value to indicate error.

The basic framework of realpath is also included in order for this
change to be tested.

PiperOrigin-RevId: 208743381