Add an fd reference count for IOContext
Currently, we are relying on the count to shared_ptrs to IOContext
objects to determine whether host close should be called. But host close
is not called in case other operations are performed. This change adds a
fd reference count to IOContext object that counts the number of file
descriptors referring to the IOContext, and calls host close when it's
the last reference.

PiperOrigin-RevId: 209788826