Return pointer to buffer in inet_ntop
Originally, a pointer to untrusted memory was being returned by
enc_untrusted_inet_ntop since it propagates the return of the ocall,
which points to a copied buffer in untrusted memory (by virtue of
the generated code). This is fixed here by properly returning a
pointer to the provided buffer upon success and nullptr on failure.

PiperOrigin-RevId: 209866160