Fix uninitialized state in EnclaveTestLauncher
This change fixes several bugs in EnclaveTestLauncher related to its
member_ and client_ pointers being uninitialized at construction time.
This caused the TearDown() and Run() methods to segfault if SetUp() was
never called, even though TearDown() was supposed to handle this
gracefully.

The Run() method was updated to return a FAILED_PRECONDITION error if
the client_ was not initialized. Although this is inconsistent with the
behavior of TearDown(), it seems less misleading.

PiperOrigin-RevId: 210751211