Remove --batch from container bazelrc
This flag was used as a workaround for a bug in Bazel where it would
time out waiting to connect to the server when run in a container.
Bazel 0.17.1 was just released which fixes that bug and allows us to
remove this workaround.
This also removes build warnings that began appearing with Bazel
0.15.0 because --batch became deprecated.

PiperOrigin-RevId: 213330604