Add docker flags for using HW enclaves
Using SGX hardware enclaves requires supporting software be installed
on the host. When running Asylo in a container, Docker defaults to not
providing access to that software on the host. Rather than duplicating
that functionality in the container, it is easier for the user to just
forward it from the host to the container.

PiperOrigin-RevId: 213675658