Move project bazelrc to new location
Starting with Bazel 0.18*, Bazel will no longer find configuration files
at tools/bazel.rc. Instead, .bazelrc at the project root now is used as
workspace-specific configuration (previously, that was considered user
configuration).

In Bazel 0.17, a bug was fixed** that prevented us from placing our
bazel configuration at .bazelrc in the project root. Since that's
no longer an issue, we can move our project configuration in
preparation of Bazel 0.18. This way our builds won't break when that is
released.

This does have the side-effect that user-specific configuration in
~/.bazelrc will be ignored until Bazel 0.18 fixes the situation. The
presence of such configuration is hopefully rare and users Bazel-savvy
enough to have it can adapt for the couple of weeks until Bazel 0.18 is
released.

* See https://github.com/bazelbuild/bazel/commit/5e5ee0d
** See https://github.com/bazelbuild/bazel/commit/20500d3

PiperOrigin-RevId: 214328115