Use copy_from_host in embed_enclaves
The copy_from_host macro is the preferred method for importing host
toolchain dependencies into a rule that is meant to be built with the
enclave toolchain. The embed_enclaves macro depends on the host
toolchain to make the enclave loader, the enclave toolchain to build the
enclaves to embed, and the enclave toolchain's objcopy to embed the
enclave binaries into the loader binary.

PiperOrigin-RevId: 214497559