Extend the protobuf printer for SGX protos
This change adds a library that creates a special proto2 printer for
printing sgx::CodeIdentity, sgx::CodeIdentityMatchSpec, and
sgx::CodeIdentityExpectation protos. This printer is different from the
standard printer in that it prints bytes fields in hex, and it prints
the ATTRIBUTE bit name for each bit that is set in an ATTRIBUTES bit
vector.

PiperOrigin-RevId: 219224736