Load enclave from file when available
Previously, support was added to load an enclave from a buffer. At that
time, even cases where a file was available were changed to use that
path.

It has now been found that the current GDB debugging approach is not
compatible with loading from a buffer. So, this changes the cases where
an enclave file is available to load the enclave from that file
directly.

GDB debugging should be fixed for these cases, but remains broken for
cases where the enclave is loaded from a buffer. That will need to be
addressed in a longer term fix.

PiperOrigin-RevId: 219542401