Rename BitVector128 to Attributes
This change renames BitVector128 to Attributes, and renames its fields
from low and high to flags and xfrm, respectively. BitVector128 was
defined as a general structure representing a 128-bit vector, but is
only used for the representation of SGX ATTRIBUTES. Textprotos and other
human-readable representations of ATTRIBUTES suffer as a result of the
more general naming.

This is considered a safe change because the new proto is still binary
compatible with the original definition.

PiperOrigin-RevId: 221129120