Support setting strict self identity expectations
This change adds a function for setting a strict self identity
expectation, which requires a match on all parts of the identity. This
function is useful for enforcing ACLs where the expected identity is
identical to the self identity (the identity of the currently-executing
enclave).

This change also adds a function for setting a strict code identity
match spec, which enforces a match on all security-sensitive parts of
the identity. This is used in the implementation of the self identity
MRENCLAVE expectation, and can be used in conjunction with
SetExpectation() to create a strict expectation on another identity.

PiperOrigin-RevId: 221379916