Refactor out InstantiateSafeBytes to bytes.h
Given that InstantiateSafeBytes is a SafeBytes util that could
be used with multiple tests that transform hex strings to
SafeBytes objects, refactored to bytes.h.

Also changed the function signature to return a StatusOr to
cover bad input. To get further information on the reason for
failure, changed SetTrivialObjectFromHexString to return a
Status instead of a Boolean.

PiperOrigin-RevId: 221474666