Update proto_format to format arbitrary messages
Previously, the proto_format library did not accept arbitrary proto
messages because the BitVector128 message could not be interpreted as a
representation of SGX ATTRIBUTES in all cases. Now that this proto has
been renamed to Attributes, and its fields are used to represent
particular SGX attribute bits, it makes sense to enable a more generic
printer interface that accepts any proto message.

PiperOrigin-RevId: 221542450