Implement pthread_cleanup_push and pop
POSIX defines pthread_cleanup_push and pthread_cleanup_pop,
allowing threads to define cleanup functions that get executed
when the thread ends or is cancelled. This change adds support
for those functions in Asylo.

Note that pthread_cancel is currently not supported in Asylo,
so the only situation in which registered cleanup functions
will execute is if a thread returns early (voluntarily).

PiperOrigin-RevId: 221693229