Change ocall failure to puts an error and abort
In the case that an ocall fails for reasons related to the bridge
between trusted and untrusted code, the current semantics for
host_calls is insufficient to explain the failure. There is no
reasonable way to recover from a failed ocall, so instead of
setting errno and returning something that can be interpreted as
an error value, the bridge will log the failure and abort.

If logging itself fails its own ocalls, then the recursive call
to LOG(FAIL) will not attempt any ocalls and simply abort.

PiperOrigin-RevId: 222277637