Provide mkdir symbol in Asylo
When the mkdir implementation was added to Asylo, it was thought that
newlib would provide the mkdir symbol and route to Asylo, as is done
with many other calls that newlib uses internally. A temporary define
was added to Asylo's stat.h so that we could test mkdir in a fashion
compatible both with newlib versions with and without this symbol.

However, it seems that, while newlib routes _mkdir calls to Asylo, it
does not provide the mkdir symbol itelf. This change removes the
temporary define and replaces it with a proper symbol that calls the
underlying implementation.

PiperOrigin-RevId: 224233121