Define substructs of SIGSTRUCT
This change introduces new structures that contain subsets of the fields
in SIGSTRUCT. Some of these structures are added for convenience of
grouping related fields together. The SigstructSigningData structure
defines the subset of fields from SIGSTRUCT that are signed. This
structure is not defined a separate structure in the Intel SDM, but it
is considered architectural.

PiperOrigin-RevId: 224253821