Fix bugs: getsockname() and its callers; add tests
getsockname() had a crash bug fixed in this change. On the host side,
the driver would call getsockname() and then marshal the resulting
sockaddr into a bridge proto and send it, where the enclave would
unmarshal it. This marshalling and unmarshalling would happen even if
the underlying call to getsockname() failed, causing uninitialized
memory to be marshalled, which would cause an abort() in the enclave
when it tried to unmarshal.

The fix is to only marshal and unmarshal the sockaddr if the call to
getsockname() succeeds.

A related bug: if FromBridgeSockaddr was passed a too-small buffer, the
converted sockaddr would overwrite memory, causing undefined results.
This change modifies FromBridgeSockaddr to follow POSIX semantics of
truncating the sockaddr rather than overwriting non-owned memory. In
places where we call FromBridgeSockaddr internally, we check-fail if it
turns out we passed to small of a buffer in. Clients who call
FromBridgeSockaddr implicitly via getsockname() will get standard POSIX
truncation semantics.

Related to the above change, there were 3 other bugs unearthed.
socket_server.cc was incorrectly passing the length of a pointer rather
than the length of the pointed-to buffer to getsockname().  Two calls to
FromBridgeSockaddr in ocalls.cc were using a bridge_sockaddr as storage
instead of a sockaddr_storage.

Finally, this change adds unit tests for getsockname(), and adds
protection against invalid pointers.

PiperOrigin-RevId: 225461184