Add methods for message limits to SecretSealer
Adds MaxMessageSize and MaxSealedMessages to SecretSealer. This allows
the user to learn the max message size they can encrypt, for chunking.
It also gives a heuristic to determine safe limits for the number of
chunks that they may encrypt.

Adds helper methods to local_secret_sealer_helpers for the
implementation in SgxLocalSecretSealer, and adds methods to AeadCryptor
to help determine the max message size and the max number of sealed
messages from the data that is available from just the cipher that will
be used.

PiperOrigin-RevId: 229983756