Change the default behavior of std::random_device
The current default behavior of std::random_device contructor is
to use a predefined seed value for a pseudo random number generator.
This is because the macro _GLIBCXX_USE_RANDOM_TRT1 is not defined
when building libstdc++.

Asylo enclaves have already supported "/dev/random" and "/dev/urandom"
for generating randomness. This commit hardcodes the macro
"_GLIBCXX_USE_RANDOM_TR1" in the install script so that
std::random_device can rely on the random source "/dev/urandom"
instead of returning always a deterministic random number sequence.
PiperOrigin-RevId: 231826487