Change ecall_restore pointer attributes
Change ecall_restore pointer attributes to "user_check" instead of "in"
or "out". SGX makes a copy of untrusted buffer on enclave heap if the
ecall pointer attribute is "in" or "out". These buffers are wiped out by
ecall_restore when restoring the whole enclave heap, which causes
segfault when sgx is freeing them before exiting the enclave. So in this
special ecall we are using "user_check" to pass the untrusted pointers
directly.
SGX enclaves have access to untrusted memory, so inside the enclave the
untrusted input could be read directly, and the output is a status that
is allocated on untrusted memory already. So we don't need a copy of
them inside the enclave, and by doing so no extra enclave info is
written to the untrusted side either.
Also, since the pointers are passed directly into the enclave, remove
"free" of the output pointer to avoid freeing the memory twice.

PiperOrigin-RevId: 233490077