Migrate cc_toolchains to *_runtime_lib
Bazel is removing old cc_toolchain.(static|dynamic)_runtime_libs
attributes and is replacing them with
cc_toolchain.(static|dynamic)_runtime_lib.

The old attribute works as follows:

1) a list of labels is read from cc_toolchain attribute

2) all these labels are analyzed and added as dependencies of the
cc_toolchain target

3) one of those labels must be selected by
CROSSTOOL.toolchain.(static|dynamic)_runtimes_filegroup

4) that label is loaded and added to actions as input

This behavior was added in times when one cc_toolchain could be
associated with multiple CROSSTOOL.toolchains. This is no longer
the case, there is 1:1 correspondence between cc_toolchain and
CROSSTOOL.toolchain now. As a result, it is always just one label
that matches, so we can skip steps 2 and 3.

This change requires Bazel 0.22.0, so that dependency is also
updated in INSTALL.md.

PiperOrigin-RevId: 234014215