Add authority config for SGX local authority
This change adds an authority config proto for SGX local assertion
authorities that encapsulates the configuration details needed by those
authorities. This config contains the local attestation domain.
Previously, this value was read from the global EnclaveConfig.

With this change, the local attestation domain is now read from the
SgxLocalAssertionAuthorityConfig proto, which is serialized and passed
to the Initialize() method of each SGX local assertion authority
library.

Updated various tests to stop setting the global EnclaveConfig.

PiperOrigin-RevId: 235659988
Change-Id: Iccd32efd549f1781109c963c32b1cedf35bc6b29