Removed Undefined Behavior in syslog.cc
Currently have undefined behavior in syslog.cc due to the over use of
the va_list args variable after the vsnprintf to calculate the size
of the buffer to allocate.

The va_list args needs to be reassigned prior to use to prevent
undefined behavior in the population of the buffer for the syslog
logging method.

PiperOrigin-RevId: 237162166