Allow user utility tcs to handle exception
Currently if an exception occurs while a signal is being handled, it is
delivered to the untrusted thread, and enters the enclave through the
bound tcs, which is not the active user utility tcs, and crashes. This
commit sets utility_thread_id every time when a thread enters the
enclave through utility tcs, and when a signal handler tries to enter
the enclave during then, it checks the thread id and enters the utility
tcs if the thread id matches the stored utility_thread_id.

PiperOrigin-RevId: 238309669