Move #include_next outside header guards
When #include_next loops back on the same file through a duplicate
path in the include paths list, it should continue to search for
the next header file. With the enabling of the Bazel flag
--incompatible_remap_main_repo, the external/com_google_asylo/asylo/...
path does not exist when building Asylo targets. However, dependencies
require references through the external/ path. Thus the CROSSTOOL
makes both path forms available. With this flag set to false (pre-0.24),
the duplicate search paths require the #include_next uses to continue
their search.

PiperOrigin-RevId: 240182299
Change-Id: Ic4296d782f0339e9ef6bd34a9b70372bf72a9977