Remove long double math functions
Newlib does not implement long double versions of math functions, so we
used the GCC builtins to implement them. However, some of the builtins
simply call the back to the function itself instead of providing an
actual implementation. This leads to an endless recursion.

Until/unless we gain an actual implementation of them, these are
removed.

PiperOrigin-RevId: 240997987
Change-Id: Ida9185aad498b6d6131c6c2d5485fc6a7b3418e6