Move type conversions to system_call library
This change moves the type conversions directory and all libraries it
contains to under system_call library. The reason for doing this is
that we expect type conversions to be used both by the system_call
and host_call libraries, but system_call library should not depend on
host_call library.

PiperOrigin-RevId: 242161094