Add PushCopy to ParameterStack
That new API would make a copy of the extent on the heap and let
the stack own it. It is suitable for cases where the lifetime of
the stack exceeds the lifetime of the object referred to by extent,
for example, in order to pass args sent across the enclave boundary
in any backend that does not allow shared access between trusted and
untrusted components.

PiperOrigin-RevId: 242763032