Keep unused bytes in the queue for fork handshake
Currently the fork key transfer starts with an Ekephandshake between the
parent and the child enclave, sending messages through a socketpair. The
parent initializes the handshake, and sends the last message during the
handshake. After the handshake, the parent sends the encrypted snapshot
key to the child. However, there are cases that when the parent finishes
the handshake, encrypts the key, and sends it through the socketpair,
the child hasn't finished the handshake yet, and in the last step of
handshake reads both the handshake message and the encrypted key from
the socketpair. This commit makes the parent enclave waits for the child
enclave ready message before sending the encrypted snapshot key.

PiperOrigin-RevId: 243437943
Change-Id: I0af8bcd0a80451b3c18a814d3c6f9906706d6b60