Implement struct types generation for primitives
This change modifies the type_conversions library to add support for
generating bridge struct definitions and conversion functions between
bridge structs and native library based structs. In cases where the
conversion functions are complex and need to be manually written, an
option to skip the conversions function generation is provided, only
generating the bridge struct definition.

PiperOrigin-RevId: 243648679
Change-Id: I45253c4111f7af83e58b1f729e57e40479944be1