Make parent fork return error if child failed
Currently parent fork returns success as long as it finishes the
handshake and sends the encrypted key. However, the child could fail to
restore the enclave if it failed to receive the key or restore the
enclave. This causes false negative test results. This commit creates a
pipe for the child to inform the parent whether it actually succeeds,
and the parent waits for the information before returning from fork.

PiperOrigin-RevId: 243673662
Change-Id: I97a98f2582cba7931bffb487777ace5b48d470a2