Move generate_test_vectors.go to testdata sub-dir
This change moves the generate_test_vectors.go binary to a testdata
subdirectory. This target only needs to be run from the WORKSPACE. It is
not needed for baseline Asylo usage and, consequently, users should not
be required to load the necessary go deps from their WORKSPACE when
using libraries from the grpc/auth package.

PiperOrigin-RevId: 243719299
Change-Id: I0af6ba885044bdede0a0f67ecec07c655ca6321d