Add proto for asymmetric signing or verifying key
This change adds a new message, AsymmetricSigningKey, that can be used
to represent either a signing key or verifying key. The primary intent
is to use this message to represent the Assertion Generator Enclave's
attestation key. A message containing the private key will be sealed by
the Assertion Generator Enclave to an asylo.SealedSecret. A message
containing the public key will be sent to Intel's PCE for certification.

The AsymmetricSigningKey message is sufficiently generic so that it can
also be used to represent other types of signing keys.

This change also updates the AsymmetricEncryptionKey message so that it
can be used to represent either an encryption key or a decryption key.
This is for consistency with the AsymmetricSigningKey message.

PiperOrigin-RevId: 243902191
Change-Id: I954fbb31b92f1a7c0e2942bbf807b979a8da2ef7