Add dependency for Intel SGX DCAP v1.0.1
This change adds Intel SGX DCAP as a dependency to Asylo. It adds a
BUILD file that provides bazel support for building the DCAP PCE
wrapper.

As part of this change, the linux SGX patch file is updated to allow
DCAP to take a dependency on some targets exposing SDK headers.

PiperOrigin-RevId: 244734687
Change-Id: I86281ac98febc81c6c65833b8f63ea7bb314877e