Implement enc_untrusted_flock() for primitives
This change implements the host call enc_untrusted_flock() in the
host_call library for the primitives interface. It provides the
enum definitions for flock operations and tests for validating the
generated conversion functions.

This change also adds the unit tests for validating the generated
conversion functions for the enum definitions TcpOptionName,
IpV6OptionName, SocketOptionName.

PiperOrigin-RevId: 244891559
Change-Id: I5278b30e7314b3399e1f665bb469e7324b60f833