Use enclave address as associated data during fork
Currently fork uses the same string as associated data for the whole
data section, same for bss/heap/thread/stack. This commit changes to use
the enclave address being encrypted/restored as the associated data, to
make sure the memory restored in the child enclave is always encrypted
from exactly the same address space in the parent enclave.

PiperOrigin-RevId: 245263448
Change-Id: I0e3868bff6e0a1dd309e1fcb3644d8aea32a7df8