Add proto definition, util library for MISCSELECT
This change adds a proto representation for the MISCSELECT bit vector.

It also defines an enum class corresponding to SECS MISCSELECT bits.
This library also provides functions for operating on a MISCSELECT bit
vector. This library is rolled into the hardware_types target, similarly
to the existing secs_attributes library.

Lastly, this change updates the proto format library for SGX protos to
format the text representation of a Miscselect proto using the names of
the MISCSELECT bits.

PiperOrigin-RevId: 246856737
Change-Id: I229fcc58b93b4504818bf3d50d1b0ff370da0fd2