Add Intel architectural enclave interface
This change adds an interface for interactions with Intel architectural
enclaves. This interface is a very lightweight wrapper around the AE
DCAP interfaces, except it uses Asylo representations of SGX
architectural structures instead of SGX-SDK-defined representations of
SGX architectural structures.

This change also adds an implementation of the interface that calls the
corresponding function in the DCAP library. This implementation is
untested because it simply invokes the DCAP library and doesn't have any
complex logic.

PiperOrigin-RevId: 247064556
Change-Id: Ibf24494945420926d42318b9fdec1d04c8722c71