Update the KeyDependencies structure
This change updates the KeyDependencies structure to always have a
fixed 8192-byte size. This is achieved by moving all of the SDM-defined
key-dependencies fields to a new structure called KeyDependenciesBase,
and defining KeyDependencies to contain KeyDependenciesBase and a
padding.

This change also modifies the renamed KeyDependenciesBase structure to
include a PKCS padding field and the KEYPOLICY field.

PiperOrigin-RevId: 247217566
Change-Id: I386b1cb27a979799adc87f144723f4b48424d727