Add GetIdentity() method to FakeEnclave
This change adds a GetIdentity() method to FakeEnclave that can be
invoked on a FakeEnclave instance to get a CodeIdentity proto
representation of its identity, which is useful for debugging in tests.
This is different than calling SelfIdentity()->identity because it does
not require the FakeEnclave::current_ pointer to be set.

PiperOrigin-RevId: 248404958
Change-Id: Ia97089d89f271f6e6bd9b517a49345d4fbc09910