Implement getpwuid
getpwuid returns a pointer to a struct `passwd`, which contains a few
char * pointers. This commit implements it by storing a global struct
`passwd` inside the enclave, along with global buffers for each field.
`getpwuid` will exit the enclave, calls `getpwuid` on the host, copies
the buffers into the enclave global buffers, and directs the pointers
in the global `passwd` struct to point to the global buffers. In the end
a pointer to the global struct is returned.

This implementation is NOT thread safe, same as this call on the host.

PiperOrigin-RevId: 249869656
Change-Id: I855f9aea1706aa0181f8b22648f739865c470a94