Sgx UntrustedCall implementation for Primitives
This change provides the implementation for making UntrustedCall in the
primitives interface for SGX backend.
UntrustedCall supports both TrustedParameterStack and
UntrustedParameterStack passed to it. If the ParameterStack passed is
on the untrusted side, UntrustedCall performs the ocall directly. If
however, the ParameterStack pointer passed to it belongs to trusted
memory, a new ParameterStack is created in the untrusted memory
and the trusted data is copied to it before invoking the ocall.

PiperOrigin-RevId: 249878507
Change-Id: Idab11c391e7a066aa7982fbc2716ce9a417844d8