Make ConvertToSockaddr return the actual addrlen
The current implementation of getaddrinfo passes the ai_addrlen from the
host directly to the enclave ai_addrlen in struct addrinfo. This may
cause crashing or even potential security issue, as the newlib size of
the struct could very well be different from the host. This commit
changes it to let the ConvertToSockaddr method determine the addrlen
based on the type of the socket.

PiperOrigin-RevId: 250403257
Change-Id: Iff28e6fbcaa214354aa120a8f86697562e6abe16