Set TLS in simulation for signal handlers
Currently signal handlers registered on the host can cause issues in SGX
simulation mode if they are accessing TLS, since if the signal arrives
while the thread is executing inside the enclave, the thread register is
still pointing to the TLS inside the enclave, and the signal handler can
not access the thread local storage on the host side.
This commit checks all registered signal handlers prior to the enclave
is created, stores them in a map, and registers a new handler for them,
which sets the thread register back to untrusted TLS if the TCS is
actively running inside the enclave, invokes the registered handler, and
then switch the thread register back.

The changes are only applied to SGX simulation mode.

PiperOrigin-RevId: 251958642
Change-Id: I767d109be14f6376f750d3fd8fc2efaaf7956868